dev/server/backend/api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "nogi", email = "ok220075@oks.iput.ac.jp" }
]
dependencies = [
    "fastapi>=0.115.12",
    "uvicorn>=0.34.0",
    "pydantic>=2.11.4",
    "pillow>=11.2.1",
    "sqlalchemy>=2.0.41",
    "python-dotenv>=1.1.0",
    "psycopg2-binary>=2.9.10",
    "google-generativeai>=0.8.5",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/api"]
